pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "networkx>=3.5",
    "numpy>=2.2.6",
    "pandas>=2.3.0",
    "psycopg2-binary>=2.9.10",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "werkzeug>=3.1.3",
]
